Neural_Information_Processing_Project_/requirements.txt
absl-py==0.8.0
alabaster==0.7.12
asn1crypto==1.0.1
astor==0.8.0
astroid==2.3.1
atomicwrites==1.3.0
attrs==19.2.0
Babel==2.7.0
backcall==0.1.0
biwrap==0.1.6
bleach==3.1.0
boltons==19.3.0
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.1
contextlib2==0.6.0.post1
cryptography==2.7
cycler==0.10.0
cytoolz==0.10.0
dask==2.5.2
debtcollector==1.22.0
decorator==4.4.0
defusedxml==0.6.0
dit==1.2.3
docutils==0.15.2
entrypoints==0.3
future==0.18.2
gast==0.3.2
grpcio==1.16.1
h5py==2.9.0
idna==2.8
imageio==2.6.0
imagesize==1.1.0
importlib-metadata==0.23
innvestigate==1.0.8
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.3
joblib==0.13.2
jsonschema==3.0.2
jupyter-client==5.3.3
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mkl-service==2.3.0
more-itertools==7.2.0
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
notebook==6.0.1
numpy==1.14.2
numpydoc==0.9.1
olefile==0.46
packaging==19.2
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
patsy==0.5.1
pbr==5.4.4
pickleshare==0.7.5
Pillow==6.2.0
pluggy==0.13.0
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
psutil==5.6.3
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyreadline==2.1
pyrsistent==0.15.4
PySocks==1.7.1
pytest==5.2.2
python-dateutil==2.8.0
python-pptx==0.6.18
pytz==2019.3
PyWavelets==1.0.3
pywin32==223
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
requests==2.22.0
rope==0.14.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==2.0.0
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==3.3.4
spyder-kernels==0.5.2
statsmodels==0.10.1
tensorboard==1.12.2
tensorflow==1.12.0
tensorflow-plot==0.3.2
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.2
toolz==0.10.0
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.2
zipp==0.6.0
